feat(dead-code): removed chat unused
